requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.5
attrs==24.2.0
azure-cognitiveservices-speech==1.41.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
decorator==4.4.2
dill==0.3.9
distro==1.9.0
exceptiongroup==1.2.2
ffmpeg-python==0.2.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
googleapis-common-protos==1.65.0
groq==0.11.0
grpcio==1.67.0
grpcio-status==1.67.0
gTTS==2.5.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
isort==5.13.2
Jinja2==3.1.4
jiter==0.6.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mccabe==0.7.0
mdurl==0.1.2
moviepy==1.0.3
narwhals==1.9.3
openai==1.52.0
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
proglog==0.1.10
proto-plus==1.24.0
protobuf==5.28.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pylint==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.39.0
tenacity==9.0.0
toml==0.10.2
tomli==2.0.2
tomlkit==0.13.2
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2

tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
